Isolate ParamsChecker from AppPage
