+++ web/src/pages/App/AppPage.jsx
@@ -1,29 +1,25 @@
 import { Box } from "@mui/material";
-import { useSnackbar } from "notistack";
 import React, { useEffect } from "react";
 import { useCookies } from "react-cookie";
 import { ErrorBoundary } from "react-error-boundary";
 import { useDispatch, useSelector } from "react-redux";
-import { Outlet, useLocation, useNavigate } from "react-router-dom";
+import { useLocation, useNavigate } from "react-router-dom";
 
 import { useSkipUntilAuthTokenIsReady } from "../../hooks/auth";
-import { useGetUserMeQuery, useTryLoginMutation } from "../../services/tcApi";
+import { useTryLoginMutation } from "../../services/tcApi";
 import { setAuthToken } from "../../slices/auth";
-import { LocationReader } from "../../utils/LocationReader";
 import { mainMaxWidth } from "../../utils/const";
-import { errorToString } from "../../utils/func";
 import { authCookieName } from "../Login/LoginPage";
 
 import { AppBar } from "./AppBar";
 import { AppFallback } from "./AppFallback";
 import { Drawer } from "./Drawer";
 import { Main } from "./Main";
+import { ParamsChecker } from "./ParamsChecker";
 
 export function App() {
   /* eslint-disable-next-line no-unused-vars */
   const [cookies, _setCookie, _removeCookie] = useCookies([authCookieName]);
-
-  const { enqueueSnackbar } = useSnackbar();
 
   const skip = useSkipUntilAuthTokenIsReady();
 
@@ -32,12 +28,6 @@ export function App() {
   const location = useLocation();
   const navigate = useNavigate();
 
-  const {
-    data: userMe,
-    error: userMeError,
-    isLoading: userMeIsLoading,
-    isFetching: userMeIsFetching,
-  } = useGetUserMeQuery(undefined, { skip });
   const [tryLogin] = useTryLoginMutation();
 
   useEffect(() => {
@@ -61,34 +51,6 @@ export function App() {
     _checkToken();
   }, [cookies, dispatch, location, navigate, skip, tryLogin]);
 
-  useEffect(() => {
-    if (!userMe || userMeIsFetching) return;
-    const params = new URLSearchParams(location.search);
-    const locationReader = new LocationReader(location);
-    if (
-      locationReader.isStatusPage() ||
-      locationReader.isTagPage() ||
-      locationReader.isPTeamPage()
-    ) {
-      if (!userMe.pteam_roles.length > 0) {
-        if (params.get("pteamId")) {
-          navigate(location.pathname);
-        }
-        return;
-      }
-      const pteamIdx = params.get("pteamId") || userMe.pteam_roles[0].pteam.pteam_id;
-      if (params.get("pteamId") !== pteamIdx) {
-        params.set("pteamId", pteamIdx);
-        navigate(location.pathname + "?" + params.toString());
-        return;
-      }
-    }
-  }, [dispatch, enqueueSnackbar, navigate, location, userMe, userMeIsFetching]);
-
-  if (skip) return <></>;
-  if (userMeError) return <>{`Cannot get UserInfo: ${errorToString(userMeError)}`}</>;
-  if (userMeIsLoading) return <>Now loading UserInfo...</>;
-
   return (
     <>
       <Box flexGrow={1}>
@@ -99,7 +61,7 @@ export function App() {
         <Box display="flex" flexDirection="row" flexGrow={1} justifyContent="center" m={1}>
           <Box display="flex" flexDirection="column" flexGrow={1} maxWidth={mainMaxWidth}>
             <ErrorBoundary FallbackComponent={AppFallback}>
-              <Outlet />
+              <ParamsChecker />
             </ErrorBoundary>
           </Box>
         </Box>
